backend/backup_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
boto3==1.34.18
azure-storage-blob==12.19.0
google-cloud-storage==2.14.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
